criacao do recurso e servico TipoEvento e adição de exception nos serviços Evento e TipoEvento

+++ BackEnd/service/src/main/java/com/basis/sge/service/servico/EventoServico.java
@@ -5,6 +5,7 @@ import com.basis.sge.service.dominio.Evento;
 import com.basis.sge.service.repositorio.EventoRepositorio;
 import com.basis.sge.service.servico.dto.EventoDTO;
 
+import com.basis.sge.service.servico.dto.TipoEventoDTO;
 import com.basis.sge.service.servico.exception.RegraNegocioException;
 import com.basis.sge.service.servico.mapper.EventoMapper;
 import lombok.RequiredArgsConstructor;
@@ -35,12 +36,15 @@ public class EventoServico {
 
     public EventoDTO obterPorId(Integer id){
         Evento evento = eventoRepositorio.findById(id)
-                .orElseThrow(() -> new RegraNegocioException("Usuario não Existe"));
+                .orElseThrow(() -> new RegraNegocioException("Evento não Existe"));
         return eventoMapper.toDto(evento);
     }
 
 
     public EventoDTO criar(EventoDTO eventoDTO) {
+        if (eventoRepositorio.existsByTitulo(eventoDTO.getTitulo())){
+            throw new RegraNegocioException("Um evento com esse titulo já existe");
+        }
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoSalvo = eventoRepositorio.save(evento);
 
@@ -48,7 +52,9 @@ public class EventoServico {
     }
 
     public EventoDTO atualizar(EventoDTO eventoDTO) {
-
+        if (eventoRepositorio.existsByTituloAndIdNot(eventoDTO.getTitulo(),eventoDTO.getId())){
+            throw new RegraNegocioException("Um evento com esse titulo já existe");
+        }
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoAtualizado = eventoRepositorio.save(evento);
 
@@ -57,6 +63,9 @@ public class EventoServico {
 
 
     public void deletar(Integer id) {
+        if(!eventoRepositorio.existsById(id)){
+            throw new RegraNegocioException("Evento Não Existe");
+        }
         eventoRepositorio.deleteById(id);
     }
